Rubocop fixes for the basic_test_class
